LNA

Implementation of the LNA method

CRMP.LNA — Type

LNA

general supertype for different LNA algorithms

CRMP.LNAR — Type

LNAR

LNA method with restart, guided rates are determined through solving a system of ODEs at each reaction time. julia-repl julia> obs = partial_observation(1.0, [2, 5], [1 0 0 ; 0 1 0], 0.05) julia> LNAR(obs, GTT(1.0, 1.0, 1.0, 1.0))`

CRMP.LNAR_death — Type

LNAR_death

LNA method with restart, closed form solution to the ODEs specifically implemented for the death process. Works the same as LNAR

CRMP.LNA_nR — Type

LNA_nR

LNA method without restart. requires fillin ψt, Gt, and zt on a grid before use ```julia-repl julia> obs = partialobservation(1.0, [2, 5], [1 0 0 ; 0 1 0], 0.05) julia> timegrid = 0.0:0.001:1.0 julia> GP = LNAnR(obs, timegrid, GTT(1.0, 1.0, 1.0, 1.0)) julia> x₀ = [1, 15, 20] julia> fill_grid!(GP, x₀)

CRMP.diff_death — Type

diff_death

specific guiding term for the death process with a diffusion guiding term

julia> obs = partial_observation(1.0, 35, 1.0, 0.05)
julia> a = 15.0
julia> GP = diff_death(obs, a, PureDeathProcess_constantrate(0.5))

CRMP.pois_death — Type

pois_death

specific guiding term for the death process with a poisson guiding term

julia> obs = julia> obs = partial_observation(1.0, 35, [nothing, 1.0], 0.05)
julia> GP = pois_death(obs, PureDeathProcess_constantrate(0.5))