Conditional process

Functions used in the simulation of guided processes

CRMP.partial_observation — Type

partial_observation{T::Union{Real, AbstractArray{Real, 1}}}

A partial observation $v~N(LX(t), ϵ*L\tilde{a}L')$ at time t with an $m \times d$-matrix L and $m \ times m$ matrix C. Use

julia> obs = partial_observation(1.0, [2, 5], [1 0 0 ; 0 1 0], 0.05)
julia> gett(obs) # returns 1.0
julia> getv(obs) # returns [2, 5]
julia> getL(obs) # returns [1 0 0 ; 0 1 0]
julia> getϵ(obs) # returns 0.05
julia> getm(obs) # returns 2
julia> getd(obs) # returns 3

CRMP.partial_observation_poisson — Type

partial_observation_poisson{T::Union{Real, AbstractArray{Real, 1}}}

A partial observation for a process with a monotone component Assume an observation v = LX(t) with Lof the form[L₁ 0 ; 0 L₂].L₂` for the monotone component, Use:

julia> obs = partial_observation_poisson(1.0, [2, 5], [ [1 0], [1] ],  0.05)
julia> gett(obs) # returns 1.0
julia> getv(obs) # returns [2, 5]
julia> getL(obs) # returns [1 0 0 ; 0 0 1]
julia> getL₁(obs)# returns [1 0] 
julia> getL₂(obs)# returns [1]
julia> getϵ(obs) # returns 0.05
julia> getm(obs) # returns 2
julia> getd(obs) # returns 3

CRMP.diffusion_guiding_term — Type

diffusion_guiding_term{T}

Guided process through array of partial observations obs, with an array $d \times d$-matrices stored in a and an original ChemicalReactionProcess P.

# Guided GTT process for one partial observation at time 1.0 using a=I
julia> partial_obs = [partial_observation(1.0, [2, 5], [1 0 0 ; 0 1 0], 0.05)]
julia> GP = diffusion_guiding_term( partial_obs , [ [1 0 0 ; 0 1 0 ; 0 0 1] ] , GTT(κ₁,κ₂,dₚ,dₘ) )

CRMP.poisson_guiding_term — Type

poisson_guiding_term{T}

Guided process through array of partial observations obs, with an array $d \times d$-matrices stored in a, an integer Y that represents the amount of poisson terms and an original ChemicalReactionProcess P.

# Guided GTT process for one partial observation at time 1.0 using a=I
julia> partial_obs = [partial_observation(1.0, [2, 5], [1 0 0 ; 0 1 0], 0.05)]
julia> GP = Guided_Process( partial_obs , [ [1 0 0 ; 0 1 0 ; 0 0 1] ] , 0, GTT(κ₁,κ₂,dₚ,dₘ) )

CRMP.filter_backward — Function

filter_backward(GP::Guided_Process)

Returns the quadruple (H,F, LaL⁻¹, LC⁻¹). All are arrays of size n that containns $H(tₖ)$, $F(tₖ)$, $(Lₖ ãₖ Lₖ)⁻¹$ and $L'( ϵ Lₖ ãₖ Lₖ)⁻¹ k=1,\dots,n$. Note, we return $H̄(tₖ)$ and $F̄(tₖ)$

CRMP.log_guiding_term — Function

log_guiding_term(info, GP::diffusion_guiding_term)

Returns a function $(ℓ, t, x) \mapsto log h̃(t,x+ξ_ℓ) - log h̃(t,x)$. info should be the quadruple (H, F, LaL⁻¹ , LC⁻¹) that results from filter_backward.

CRMP.log_guiding_term_poisson_only — Function

log_guiding_term_poisson_only(GPP::poisson_guiding_term)

Returns the log-guiding term as a function of (ℓ,t,x) for only poisson components

CRMP.guiding_term — Function

guiding_term(info, GP::T) where {T<:Guided_Process}

exp of the log guiding terms

CRMP.condition_reaction — Function

condition_reaction(ℓ::reaction{S}, xT::Union{S, Array{S,1}}, T::Real, guiding_term::Function) where {S<:Real}

Input is a reaction ℓ, a desired endpoint xT, end time T and a guiding term that is a function of (ℓ, t, x) and returns $λ̃(t, x+ℓ.ξ)/λ̃(t,x)$. Returns a reaction with the same difference vector ξ, but with the conditioned rate specified in the paper.

CRMP.condition_process — Function

condition_process(P::ChemicalReactionProcess{S}, guiding_term::Function) where {S<:Real}

Returns a new ChemicalReactionProcess{S} where all reaction in the network are conditioned using condition_reaction.