Forward simulation

Forward simulation of chemical reaction processes in various settings

CRMP.next_jump — Function

next_jump(::constant_rate, t::Real, x::Union{T, Array{T,1}}, P::ChemicalReactionProcess{T}) where {T<:Real}

If the process is of constant (in time) rate, simulates exponential jump times for all reactions and returns the reaction with the shortest jump time along with the respective time.

CRMP.simulate_forward — Function

simulate_forward(::constant_rate, x₀::Union{T1, Array{T1,1}}, T::T2, P::ChemicalReactionProcess{T1}) where {T1<:Real, T2<:Real}

Forward simulation starting from x₀, untime time 'T' of the process specified in 'P' when jump rates are constant (in time).

simulate_forward(::conditional, x₀::Union{T1, Array{T1,1}}, xT, T::T2, P::ChemicalReactionProcess{T1}, d²::Function, η) where {T1<:Real, T2<:Real}

Forward simulation of the guided process specified in 'P', starting from 'x₀' and with desired end state '(T, xT)'.

CRMP.conditional — Type

conditional <: method

For specifying that we want to simulate from the guided process