Likelihood

Functions for likelihood computation

CRMP.loglikelihood — Function

loglikelihood(tt, xx, GP::Guided_Process, info)

Specifically for a guiding term exp(-0.5 x' H(t)x + F(t)'x)

CRMP.loglikelihood_general_1obs — Function

loglikelihood_general_1obs(tt,xx, GP::T, info) where {T<:Guided_Process}

General log-likelihood computation, only dependent on guiding term, main method used

CRMP.likelihood_general_1obs — Function

likelihood_general_1obs(tt,xx, GP::T, info) where {T<:Guided_Process}

General likelihood computation, only dependent on guiding term, main method used

CRMP.iscorrect_1obs — Function

iscorrect_1obs(GP::T, tt, xx) where {T<:Guided_Process}

Checks whether LX(T) = v, returns a boolean.