Networks

Various function for setting up chemical reaction networks and several well-known networks

CRMP.reaction — Type

reaction{T<:Real}

A reaction, specified through a rate function λ of time and space and a difference vector ξ. For example, a unit rate poisson has one reaction, specified as

julia> plus1 = reaction((t,x) -> x, 1)

CRMP.ChemicalReactionNetwork — Type

ChemicalReactionNetwork{T<:Real}

A chemical reaction network is set-up as a tuple of a vector of species 𝒮 with Strings as input and a vector of reaction{T}s. I still need some way to distinct between process and network

CRMP.ChemicalReactionProcess — Type

ChemicalReactionProcess{T<:Real}

A chemical reaction process is set-up as a tuple of a vector of species 𝒮 with Strings as input and a vector of reaction{T}s. I still need some way to distinct between process and network For example, a poisson process is set-up as follows

julia> plus1 = reaction((t,x) -> x, 1)
julia> PoissonProcecss = ChemicalReactionProcess(["Counts"], [plus1])

Alternatively, if there is just one reaction, or species, one could omit the Array.

CRMP.nr_species — Function

nr_species(P::ChemicalReactionProcess)

Returns how much species a ChemicalReactionProcess contains

CRMP.nr_reactions — Function

nr_reactions(P::ChemicalReactionProcess)

Returns how much reactions a ChemicalReactionProcess contains

CRMP.PoissonProcess_constantrate — Function

PoissonProcess_constantrate(rate::Real)

Returns a ChemicalReactionProcess for the Poisson process with constant rate rate.

Missing docstring.

Missing docstring for PureDeathProcess_constantrate. Check Documenter's build log for details.

CRMP.BirthDeathProcess — Function

BirthDeathProcess(birth_rate::Real , death_rate::Real)

Returns a ChemicalReactionProcess for the birth-death process with parameters birth_rate and death_rate

CRMP.GTT — Function

GTT(κ₁::T ,κ₂::T, dₘ::T, dₚ::T) where {T<:Real}

Returns a ChemicalReactionProcess for Gene transcriptiona and translation as described in section 2.1.something of Anderson & Kurtz with

transcription rate κ₁

translation rate κ₂

degradation rate of mRNA dₘ

degradation rate of protein dₚ

GTT(κ₁::T ,κ₂::T, κ₃::T, dₘ::T, dₚ::T, dD::T) where {T<:Real}

Similar to GTT but with the inclusion of dimerization with dimerization rate κ₃ and degradation rate of dimer dD

CRMP.Schlogl — Function

Schlogl(κ₁::T,κ₂::T,κ₃::T,κ₄:T) where {T<:Real}

Schlogl model (#X, #A,#B).

Missing docstring.

Missing docstring for AR. Check Documenter's build log for details.

CRMP.viral_infection — Function

viral_infection(κ₁::T, κ₂::T, κ₃::T, κ₄::T, κ₅::T, κ₆::T) where {T<:Real}

ChemicalReactionProcess for viral infection, See e.g. section 2.1.2 of Anderson & Kurtz

CRMP.enzyme_kinetics — Function

enzyme_kinetics(κ₁::T, κ₂::T, κ₃::T) where {T<:Real}

See e.g. section 2.1.3 of Anderson & Kurtz

CRMP.lotka_volterra — Function

lotka_volterra(κ₁, κ₂, κ₃)

well known, pred -> ∅ , prey -> 2prey and pred + prey = 2pred