Conditional process

Function used in the simulation of guided processes

CRMP.method — Type

method

supertype for simulation methods, might find another name for it later

CRMP.constant_rate — Type

constant_rate <: method

used when the rates of the reactions in the network are constant in time

CRMP.decreasing_rate — Type

decreasing_rate <: method

used when the rates of the reactions in the network are decreasing in time

CRMP.increasing_rate — Type

increasing_rate <: method

used when the rates of the reactions in the network are increasing in time

CRMP.gettime — Function

gettime(::constant_rate, ℓ::reaction{T}, t::Real, x::Union{T, Array{T,1}}) where {T<:Real}

Returns an Exp(ℓ.λ(t,x))-random variable, provided ''ℓ.λ(t,x)>0'. Else returns 1e10

gettime(::decreasing_rate, ℓ::reaction, t, x::Union{T, Array{T,1}}) where {T<:Real}

Returns the reaction time of reaction ℓ if ℓ.λ is decreasing in time. Uses a thinning algorithm with upper bound 'ℓ.λ(t,x)'

gettime(::increasing_rate, ℓ::reaction{T}, t, x::Union{T, Array{T,1}}, setδ::Function) where {T<:Real}

Returns the reaction time of reaction ℓ if ℓ.λ is increasing in time. Uses a thinning algorithm on compact intervals [t, t+δ] where δ is obtained from 'setδ' and time by δ is moved if the reaction time is not below δ.